feat: refactor proposer to be in main folder
