+++ nightfall-proposer/package.json
@@ -12,7 +12,7 @@
   "dependencies": {
     "async-mutex": "^0.3.2",
     "body-parser": "^1.19.2",
-    "common-files": "file:../../../common-files",
+    "common-files": "file:../common-files",
     "config": "^3.3.1",
     "cors": "^2.8.5",
     "express": "^4.17.3",
