+++ nightfall-proposer/src/routes/proposer.mjs
@@ -4,7 +4,7 @@ Some transactions are so simple that, we don't split out a separate service
 module but handle the entire request here.
 */
 import express from 'express';
-import logger from '../../../../../common-files/utils/logger.mjs';
+import logger from '../../../common-files/utils/logger.mjs';
 import { nf3SendOffchainTransaction } from '../nf3-wrapper.mjs';
 
 const router = express.Router();
