+++ wallet/.env
@@ -1,6 +1,6 @@
 SKIP_PREFLIGHT_CHECK=true
-PROPOSER_API_URL=https://proposer.polygon-nightfall.technology
 LOCAL_API_URL=http://localhost:8081
-PROPOSER_WS_URL=wss://kpg1x6a0vj.execute-api.eu-west-1.amazonaws.com/production
 LOCAL_WS_URL=ws://localhost:8082
 LOCAL_UTIL_API_URL=http://localhost:8087
+PROPOSER_API_URL=https://proposer.testnet.polygon-nightfall.technology
+PROPOSER_WS_URL=wss://pdgpiqavbl.execute-api.us-east-2.amazonaws.com/testnet
